Added AudioHandler and options menu functionality.
AudioHandler class now handles audio with audio implemented in TextButton class. Game now has options property which contains a dictionary containing options that are meant to be able to be changed during runtime unlike config which is intended to be static from runtime.

There was a significant flaw with my stack system as state enter and state leave methods were not being called as expected. The logic has been fixed now. (I hope)