Added GUI Factory class
GUI Factory class added which now creates and returns GUI instances.

GUI Factory instance is a required attribute of any StateBase object.

This means that StateBase and its child classes are no longer dependent on the GUI classes themselves with their only dependences being pygame itself, StateBase if an inherited class, and the GUIFactory which is passed as an argument instead of being a direct dependence.